requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
grpcio==1.62.2
googleapis-common-protos==1.63.0
pydantic==1.10.14
python-dotenv==1.0.1
